requirements.txt
# WhisperX Dependencies for A100
whisperx @ git+https://github.com/m-bain/whisperx.git
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.28.0
